+++ packages/api/src/procedures/jwt-authed.ts
@@ -1,4 +1,0 @@
-import { isJWTAuthed } from "../middlewares/isJWTAuthed";
-import { publicProcedure } from "./public";
-
-export const jwtAuthedProcedure = publicProcedure.use(isJWTAuthed);